More SSR, add antd library and use it a bit
