+++ client/src/app/registerparcel/steptwo/components/PeopleInfoForm.tsx
@@ -1,10 +1,13 @@
-import styles from "../page.module.css"
+import styles from "../page.module.css";
 
 import { FormInput } from "colorblind/shared/components/FormInput";
 import { ColorblindPhoneInput } from "colorblind/shared/components/PhoneInput";
-import { TerminalPicker } from "colorblind/shared/components/TerminalPicker";
-import { PersonInfo, TerminalDetails } from "colorblind/shared/lib/models/models";
+import {
+  PersonInfo,
+  TerminalDetails,
+} from "colorblind/shared/lib/models/models";
 import { useFormik } from "formik";
+import { Select } from "antd";
 
 interface Props {
   onSubmit: (data: PeopleInfo) => void;
@@ -17,7 +20,11 @@ export interface PeopleInfo {
   receiverDeliveryInfo?: PersonInfo;
 }
 
-export const PeopleInfoForm = ({ defaultValue, onSubmit, terminals }: Props) => {
+export const PeopleInfoForm = ({
+  defaultValue,
+  onSubmit,
+  terminals,
+}: Props) => {
   const {
     values: form,
     handleChange,
@@ -28,10 +35,23 @@ export const PeopleInfoForm = ({ defaultValue, onSubmit, terminals }: Props) => 
     onSubmit: onSubmit,
   });
 
+  const terminalOptions = terminals.map((terminal) => ({
+    value: terminal.id,
+    label: terminal.address,
+  }));
+
+  const onChangeFrom = (id: string) => {
+    setFieldValue("senderDeliveryInfo.terminalId", id);
+  };
+
+  const onChangeTo = (id: string) => {
+    setFieldValue("receiverDeliveryInfo.terminalId", id);
+  };
+
   const handleSubmitWithValidation = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     if (!form.senderDeliveryInfo?.terminalId) {
-      alert("Please select terminal")
+      alert("Please select terminal");
       return;
     }
     if (!form.receiverDeliveryInfo?.terminalId) {
@@ -72,12 +92,11 @@ export const PeopleInfoForm = ({ defaultValue, onSubmit, terminals }: Props) => 
           onChange={handleChange}
         />
         <label>Terminal</label>
-        <TerminalPicker
-          nonOption="Select terminal"
-          terminals={terminals}
-          onSubmit={(id) => {
-            setFieldValue("senderDeliveryInfo.terminalId", id);
-          }} />
+        <Select
+          options={terminalOptions}
+          onChange={onChangeFrom}
+          placeholder="Please select"
+        />
       </div>
       <div className={styles.receiver}>
         <h2>Receiver</h2>
@@ -107,12 +126,11 @@ export const PeopleInfoForm = ({ defaultValue, onSubmit, terminals }: Props) => 
           country={"lt"}
         />
         <label>Terminal</label>
-        <TerminalPicker
-          nonOption="Select terminal"
-          terminals={terminals}
-          onSubmit={(id) => {
-            setFieldValue("receiverDeliveryInfo.terminalId", id);
-          }} />
+        <Select
+          options={terminalOptions}
+          onChange={onChangeTo}
+          placeholder="Please select"
+        />
       </div>
       <button type="submit">Next</button>
     </form>
